Improve error messages

+++ app/src/main/java/com/edipo/uni7java/remote/WeatherRsp.java
@@ -9,7 +9,7 @@ public class WeatherRsp {
 
     public static class Main {
 
-        public int temp;
+        public float temp;
 
     }
